Convert standalone migrator into second partial migrator
